requirements.txt
mysql-connector-python
pandas
sqlalchemy
flask
flask-cors